requirements.txt
numpy>=1.22
pandas>=1.4
matplotlib>=3.5
seaborn>=0.11
scikit-learn>=1.1
scipy>=1.8
hdbscan>=0.8.28
biopython>=1.79
umap-learn>=0.5.3
psutil>=5.9
requests>=2.28
tqdm>=4.64
